feat: implement comments

+++ Src/Application/DTOs/SalvarComentarioDTO.cs
@@ -4,5 +4,5 @@ public class SalvarComentarioDTO
 {
     public int IdJogo { get; set; }
     public string Texto { get; set; } = null!;
-    public int Nota { get; set; }
+    public short Nota { get; set; }
 }
